refact: Extract JUnit 4 support module
