+++ core/src/main/resources/META-INF/services/io.github.cdiunit.internal.DiscoveryExtension
@@ -7,7 +7,6 @@ io.github.cdiunit.internal.easymock.EasyMockDiscoveryExtension
 io.github.cdiunit.internal.IgnoredClassesDiscoveryExtension
 io.github.cdiunit.internal.CdiDiscoveryExtension
 io.github.cdiunit.internal.CdiUnitDiscoveryExtension
-io.github.cdiunit.junit4.internal.InvalidRuleFieldUsageDiscoveryExtension
 io.github.cdiunit.internal.activatescopes.ActivateScopesDiscoveryExtension
 io.github.cdiunit.internal.naming.NamingDiscoveryExtension
 io.github.cdiunit.internal.events.EventsForwardingDiscoveryExtension
