+++ junit4/src/main/java/io/github/cdiunit/junit4/internal/InvalidRuleFieldUsageDiscoveryExtension.java
