+++ junit4/src/main/java/io/github/cdiunit/junit4/package-info.java
