+++ junit4/src/main/resources/META-INF/io.github.cdiunit-archive
@@ -1,0 +1,5 @@
+# CDI Unit INTERNAL use only!
+#
+# marker file for DefaultBeanArchiveScanner
+#
+# CDI Unit INTERNAL use only!
