+++ junit4/src/main/resources/META-INF/services/io.github.cdiunit.internal.DiscoveryExtension
@@ -1,0 +1,1 @@
+io.github.cdiunit.junit4.internal.InvalidRuleFieldUsageDiscoveryExtension
