+++ junit4/src/test/java/io/github/cdiunit/junit4/tests/BaseTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.cdiunit.junit4;
+package io.github.cdiunit.junit4.tests;
 
 import jakarta.enterprise.inject.spi.BeanManager;
 import jakarta.inject.Inject;
