+++ junit4/src/test/java/io/github/cdiunit/junit4/tests/TestBasicFeatures.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.cdiunit.junit4;
+package io.github.cdiunit.junit4.tests;
 
 import java.lang.annotation.Annotation;
 import java.util.ArrayList;
@@ -39,6 +39,8 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 
+import io.github.cdiunit.junit4.CdiJUnit;
+import io.github.cdiunit.junit4.CdiRunner;
 import io.github.cdiunit.*;
 import io.github.cdiunit.test.beans.*;
