+++ pom.xml
@@ -1,4 +1,5 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+				 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
 	<parent>
@@ -30,8 +31,8 @@
 		<module>support-servlet6</module>
 		<module>core</module>
 		<module>junit4</module>
-		<module>deltaspike</module>
 		<module>junit5</module>
+		<module>deltaspike</module>
 		<module>testng</module>
 		<module>spock</module>
 		<module>integration-tests</module>
@@ -50,8 +51,8 @@
 					<target>
 						<delete includeEmptyDirs="true">
 							<dirset dir=".">
-								<include name="*spotless-groovy*" />
-								<include name="*spotless-groovy*/**" />
+								<include name="*spotless-groovy*"/>
+								<include name="*spotless-groovy*/**"/>
 							</dirset>
 						</delete>
 					</target>
